+++ packages/core/src/utilities/genericMetadataProvider.ts
@@ -6,20 +6,33 @@ let state: Record<string, any> = {}; // Calibrated pixel spacing per imageId
  */
 const metadataProvider = {
   /**
-   * Adds metadata for an imageId.
+   * Adds a cloned copy of the metadata for an imageId as the given type.
+   * Note that this will strip out any functions or other non-cloneables.
+   *
    * @param imageId - the imageId for the metadata to store
    * @param payload - the payload
    */
-  add: (imageId: string, payload: any): void => {
+  add: (imageId: string, payload: { metadata: any; type: string }): void => {
+    metadataProvider.addRaw(imageId, {
+      ...payload,
+      metadata: structuredClone(payload.metadata),
+    });
+  },
+
+  /**
+   * Adds a raw metadata instances for an imageId.  This allows preserving
+   * class inheritance values and member functions/proxy instances, but runs
+   * the risk that the raw object can be modified through side affects.
+   */
+  addRaw: (imageId: string, payload: { metadata: any; type: string }) => {
     const type = payload.type;
 
     if (!state[imageId]) {
       state[imageId] = {};
     }
 
-    // Use the raw metadata, or create a deep copy of payload.metadata
-    state[imageId][type] =
-      payload.rawMetadata ?? structuredClone(payload.metadata);
+    // Use the raw metadata here
+    state[imageId][type] = payload.metadata;
   },
 
   /**