+++ packages/tools/examples/dynamicPetCt/index.ts
@@ -366,9 +366,7 @@ function setUpToolGroups() {
   // volume to use for the WindowLevelTool for the fusion viewports
   ctToolGroup.addTool(WindowLevelTool.toolName);
   ptToolGroup.addTool(WindowLevelTool.toolName);
-  fusionToolGroup.addTool(WindowLevelTool.toolName, {
-    volumeId: ptVolumeId,
-  });
+  fusionToolGroup.addTool(WindowLevelTool.toolName);
 
   [ctToolGroup, ptToolGroup, fusionToolGroup].forEach((toolGroup) => {
     toolGroup.setToolActive(WindowLevelTool.toolName, {
@@ -437,6 +435,7 @@ function setUpSynchronizers() {
   const coronalCameraSynchronizerId = 'CORONAL_CAMERA_SYNCHRONIZER_ID';
   const ctVoiSynchronizerId = 'CT_VOI_SYNCHRONIZER_ID';
   const ptVoiSynchronizerId = 'PT_VOI_SYNCHRONIZER_ID';
+  const fusionVoiSynchronizerId = 'FUSION_VOI_SYNCHRONIZER_ID';
 
   const axialCameraPositionSynchronizer = createCameraPositionSynchronizer(
     axialCameraSynchronizerId
@@ -447,8 +446,18 @@ function setUpSynchronizers() {
   const coronalCameraPositionSynchronizer = createCameraPositionSynchronizer(
     coronalCameraSynchronizerId
   );
-  const ctVoiSynchronizer = createVOISynchronizer(ctVoiSynchronizerId);
-  const ptVoiSynchronizer = createVOISynchronizer(ptVoiSynchronizerId);
+  const ctVoiSynchronizer = createVOISynchronizer(ctVoiSynchronizerId, {
+    syncInvertState: false,
+    syncColormap: false,
+  });
+  const ptVoiSynchronizer = createVOISynchronizer(ptVoiSynchronizerId, {
+    syncInvertState: false,
+    syncColormap: false,
+  });
+  const fusionVoiSynchronizer = createVOISynchronizer(fusionVoiSynchronizerId, {
+    syncInvertState: false,
+    syncColormap: false,
+  });
 
   // Add viewports to camera synchronizers
   [
@@ -494,27 +503,30 @@ function setUpSynchronizers() {
     });
   });
   [
-    viewportIds.FUSION.AXIAL,
-    viewportIds.FUSION.SAGITTAL,
-    viewportIds.FUSION.CORONAL,
+    viewportIds.PT.AXIAL,
+    viewportIds.PT.SAGITTAL,
+    viewportIds.PT.CORONAL,
+    viewportIds.PETMIP.CORONAL,
   ].forEach((viewportId) => {
-    // In this example, the fusion viewports are only targets for CT VOI
-    // synchronization, not sources
-    ctVoiSynchronizer.addTarget({
+    ptVoiSynchronizer.add({
       renderingEngineId,
       viewportId,
     });
   });
   [
-    viewportIds.PT.AXIAL,
-    viewportIds.PT.SAGITTAL,
-    viewportIds.PT.CORONAL,
     viewportIds.FUSION.AXIAL,
     viewportIds.FUSION.SAGITTAL,
     viewportIds.FUSION.CORONAL,
-    viewportIds.PETMIP.CORONAL,
   ].forEach((viewportId) => {
-    ptVoiSynchronizer.add({
+    fusionVoiSynchronizer.add({
+      renderingEngineId,
+      viewportId,
+    });
+    ctVoiSynchronizer.addTarget({
+      renderingEngineId,
+      viewportId,
+    });
+    ptVoiSynchronizer.addTarget({
       renderingEngineId,
       viewportId,
     });
@@ -523,22 +535,22 @@ function setUpSynchronizers() {
 
 async function setUpDisplay() {
   const { metaDataManager } = cornerstoneDICOMImageLoader.wadors;
-  const wadoRsRoot = 'https://d33do7qe4w26qo.cloudfront.net/dicomweb';
+  const wadoRsRoot = 'https://d3t6nz73ql33tx.cloudfront.net/dicomweb';
   const StudyInstanceUID =
-    '1.3.6.1.4.1.12842.1.1.14.3.20220915.105557.468.2963630849';
+    '1.3.6.1.4.1.14519.5.2.1.7009.2403.334240657131972136850343327463';
 
   // Get Cornerstone imageIds and fetch metadata into RAM
   let ctImageIds = await createImageIdsAndCacheMetaData({
     StudyInstanceUID,
     SeriesInstanceUID:
-      '1.3.6.1.4.1.12842.1.1.14.4.20220915.121025.435.2500855592',
+      '1.3.6.1.4.1.14519.5.2.1.7009.2403.226151125820845824875394858561',
     wadoRsRoot,
   });
 
   const ptImageIds = await createImageIdsAndCacheMetaData({
     StudyInstanceUID,
     SeriesInstanceUID:
-      '1.3.6.1.4.1.12842.1.1.22.4.20220915.124758.560.4125514885',
+      '1.3.6.1.4.1.14519.5.2.1.7009.2403.879445243400782656317561081015',
     wadoRsRoot,
   });