+++ packages/tools/examples/petCt/index.ts
@@ -61,11 +61,13 @@ const sagittalCameraSynchronizerId = 'SAGITTAL_CAMERA_SYNCHRONIZER_ID';
 const coronalCameraSynchronizerId = 'CORONAL_CAMERA_SYNCHRONIZER_ID';
 const ctVoiSynchronizerId = 'CT_VOI_SYNCHRONIZER_ID';
 const ptVoiSynchronizerId = 'PT_VOI_SYNCHRONIZER_ID';
+const fusionVoiSynchronizerId = 'FUSION_VOI_SYNCHRONIZER_ID';
 let axialCameraPositionSynchronizer;
 let sagittalCameraPositionSynchronizer;
 let coronalCameraPositionSynchronizer;
 let ctVoiSynchronizer;
 let ptVoiSynchronizer;
+let fusionVoiSynchronizer;
 let mipToolGroup;
 const viewportIds = {
   CT: { AXIAL: 'CT_AXIAL', SAGITTAL: 'CT_SAGITTAL', CORONAL: 'CT_CORONAL' },
@@ -375,9 +377,7 @@ function setUpToolGroups() {
   // volume to use for the WindowLevelTool for the fusion viewports
   ctToolGroup.addTool(WindowLevelTool.toolName);
   ptToolGroup.addTool(WindowLevelTool.toolName);
-  fusionToolGroup.addTool(WindowLevelTool.toolName, {
-    volumeId: ptVolumeId,
-  });
+  fusionToolGroup.addTool(WindowLevelTool.toolName);
 
   [ctToolGroup, ptToolGroup, fusionToolGroup].forEach((toolGroup) => {
     toolGroup.setToolActive(WindowLevelTool.toolName, {
@@ -439,8 +439,18 @@ function setUpSynchronizers() {
   coronalCameraPositionSynchronizer = createCameraPositionSynchronizer(
     coronalCameraSynchronizerId
   );
-  ctVoiSynchronizer = createVOISynchronizer(ctVoiSynchronizerId);
-  ptVoiSynchronizer = createVOISynchronizer(ptVoiSynchronizerId);
+  ctVoiSynchronizer = createVOISynchronizer(ctVoiSynchronizerId, {
+    syncInvertState: false,
+    syncColormap: false,
+  });
+  ptVoiSynchronizer = createVOISynchronizer(ptVoiSynchronizerId, {
+    syncInvertState: false,
+    syncColormap: false,
+  });
+  fusionVoiSynchronizer = createVOISynchronizer(fusionVoiSynchronizerId, {
+    syncInvertState: false,
+    syncColormap: false,
+  });
   // Add viewports to camera synchronizers
   [
     viewportIds.CT.AXIAL,
@@ -485,27 +495,30 @@ function setUpSynchronizers() {
     });
   });
   [
-    viewportIds.FUSION.AXIAL,
-    viewportIds.FUSION.SAGITTAL,
-    viewportIds.FUSION.CORONAL,
+    viewportIds.PT.AXIAL,
+    viewportIds.PT.SAGITTAL,
+    viewportIds.PT.CORONAL,
+    viewportIds.PETMIP.CORONAL,
   ].forEach((viewportId) => {
-    // In this example, the fusion viewports are only targets for CT VOI
-    // synchronization, not sources
-    ctVoiSynchronizer.addTarget({
+    ptVoiSynchronizer.add({
       renderingEngineId,
       viewportId,
     });
   });
   [
-    viewportIds.PT.AXIAL,
-    viewportIds.PT.SAGITTAL,
-    viewportIds.PT.CORONAL,
     viewportIds.FUSION.AXIAL,
     viewportIds.FUSION.SAGITTAL,
     viewportIds.FUSION.CORONAL,
-    viewportIds.PETMIP.CORONAL,
   ].forEach((viewportId) => {
-    ptVoiSynchronizer.add({
+    fusionVoiSynchronizer.add({
+      renderingEngineId,
+      viewportId,
+    });
+    ctVoiSynchronizer.addTarget({
+      renderingEngineId,
+      viewportId,
+    });
+    ptVoiSynchronizer.addTarget({
       renderingEngineId,
       viewportId,
     });