+++ playwright.config.ts
@@ -38,6 +38,6 @@ export default defineConfig({
     command: 'yarn build-and-serve-static-examples',
     url: 'http://localhost:3000',
     reuseExistingServer: !process.env.CI,
-    timeout: 120 * 1000,
+    timeout: 150 * 1000,
   },
 });