+++ packages/adapters/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.83.1](https://github.com/dcmjs-org/dcmjs/compare/v1.83.0...v1.83.1) (2024-07-27)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [1.83.0](https://github.com/dcmjs-org/dcmjs/compare/v1.82.7...v1.83.0) (2024-07-24)
+
+### Features
+
+-   **pmap:** added parametric map adapter ([#1382](https://github.com/dcmjs-org/dcmjs/issues/1382)) ([4b5705d](https://github.com/dcmjs-org/dcmjs/commit/4b5705d937c202fb9f84c5aac2158e9577aa5aa4))
+
+## [1.82.7](https://github.com/dcmjs-org/dcmjs/compare/v1.82.6...v1.82.7) (2024-07-24)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.82.6](https://github.com/dcmjs-org/dcmjs/compare/v1.82.5...v1.82.6) (2024-07-23)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.82.5](https://github.com/dcmjs-org/dcmjs/compare/v1.82.4...v1.82.5) (2024-07-22)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.82.4](https://github.com/dcmjs-org/dcmjs/compare/v1.82.3...v1.82.4) (2024-07-10)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.82.3](https://github.com/dcmjs-org/dcmjs/compare/v1.82.2...v1.82.3) (2024-07-10)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.82.2](https://github.com/dcmjs-org/dcmjs/compare/v1.82.1...v1.82.2) (2024-07-10)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.82.1](https://github.com/dcmjs-org/dcmjs/compare/v1.82.0...v1.82.1) (2024-07-10)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [1.82.0](https://github.com/dcmjs-org/dcmjs/compare/v1.81.6...v1.82.0) (2024-07-05)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.81.6](https://github.com/dcmjs-org/dcmjs/compare/v1.81.5...v1.81.6) (2024-07-05)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.81.5](https://github.com/dcmjs-org/dcmjs/compare/v1.81.4...v1.81.5) (2024-07-05)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.81.4](https://github.com/dcmjs-org/dcmjs/compare/v1.81.3...v1.81.4) (2024-07-05)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
 ## [1.81.3](https://github.com/dcmjs-org/dcmjs/compare/v1.81.2...v1.81.3) (2024-07-02)
 
 **Note:** Version bump only for package @cornerstonejs/adapters